Create standalone classes with getList methods
