+++ Sources/LightPortal/Lists/AbstractPageList.php
@@ -32,8 +32,6 @@ abstract class AbstractPageList implements PageListInterface
 
 	abstract public function showAll();
 
-	abstract public function getList(): array;
-
 	abstract public function getAll(int $start, int $items_per_page, string $sort): array;
 
 	protected function getPreparedResults(array $rows = []): array
