+++ Sources/LightPortal/Lists/CategoryList.php
@@ -1,0 +1,50 @@
+<?php declare(strict_types=1);
+
+/**
+ * CategoryList.php
+ *
+ * @package Light Portal
+ * @link https://dragomano.ru/mods/light-portal
+ * @author Bugo <bugo@dragomano.ru>
+ * @copyright 2019-2023 Bugo
+ * @license https://spdx.org/licenses/GPL-3.0-or-later.html GPL-3.0-or-later
+ *
+ * @version 2.1
+ */
+
+namespace Bugo\LightPortal\Lists;
+
+use Bugo\LightPortal\Helper;
+
+if (! defined('SMF'))
+	die('No direct access...');
+
+final class CategoryList
+{
+	use Helper;
+
+	public function getList(): array
+	{
+		$request = $this->smcFunc['db_query']('', /** @lang text */ '
+			SELECT category_id, name, description, priority
+			FROM {db_prefix}lp_categories
+			ORDER BY priority',
+			[]
+		);
+
+		$items = [0 => ['name' => $this->txt['lp_no_category']]];
+		while ($row = $this->smcFunc['db_fetch_assoc']($request)) {
+			$items[$row['category_id']] = [
+				'id'       => $row['category_id'],
+				'name'     => $row['name'],
+				'desc'     => $row['description'],
+				'priority' => $row['priority']
+			];
+		}
+
+		$this->smcFunc['db_free_result']($request);
+		$this->context['lp_num_queries']++;
+
+		return $items;
+	}
+}
